feat: move swap widget components to @swapkit/ui/react
